:sparkles: change App to tsx
